requirements.txt
# CoincallTrader Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.31.0          # HTTP client for API calls
python-dotenv>=1.0.0      # Environment variable management
flask>=3.0.0              # Web dashboard
tomli>=2.0.0              # TOML parser (Python <3.11 fallback for tomllib)

websockets>=12.0          # WebSocket client (tick recorder)
numpy>=1.24               # Tick recorder — array ops in snapshotter
pandas>=2.0               # Tick recorder — DataFrame / parquet writes
pyarrow>=14.0             # Tick recorder — parquet backend
python-dotenv>=1.0        # Tick recorder — load .env in recorder.py

# Development dependencies (optional)
# pytest>=7.0.0           # Testing framework
# pytest-asyncio>=0.21.0  # Async test support
